requirements.txt
celery==5.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.5.0
django-phonenumber-field==5.0.0
django-shopping-cart==0.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
flower==1.0.0
phonenumbers==8.12.19
Pillow==8.3.2
PyJWT==1.7.1
python-decouple==3.5
redis==3.5.3
requests==2.26.0
psycopg2>=2.7,<2.8 --no-binary psycopg2